Moved tests back to one dir
