--- test/e2eday2/main_test.go
@@ -1,11 +1,0 @@
-package e2eday2
-
-import (
-	"testing"
-
-	f "github.com/operator-framework/operator-sdk/pkg/test"
-)
-
-func TestMain(m *testing.M) {
-	f.MainEntry(m)
-}
